fix(service): 修正数据库类型

+++ service/packages/db/tables/subTasks.ts
@@ -28,7 +28,7 @@ export const SubTasks = sequelize.define<Model<Table, Table>, Table>(
     completed_date_time: DataTypes.BIGINT,
     last_edit_time: DataTypes.BIGINT,
     parent_id: DataTypes.BIGINT,
-    owner: DataTypes.BIGINT
+    owner: DataTypes.INTEGER
   },
   {
     timestamps: false,
